custom-tooling/bws_secret_injector/requirements.txt
PyJWT==2.8.0
cryptography==42.0.8
requests==2.32.3
PyNaCl==1.5.0